LoginProject/requirements.txt
alembic==1.6.2
apifairy==0.6.2
apispec==4.4.1
appdirs==1.4.4
black==21.5b1
blinker==1.4
click==8.0.0
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.0
greenlet==1.1.0
gunicorn==20.1.0
idna==3.1
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
mypy-extensions==0.4.3
password-validator==1.0
pathspec==0.8.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
webargs==8.0.0
Werkzeug==2.0.0
WTForms==2.3.3
